+++ buildpacks/ruby/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.0] - 2023-09-26
+
 ### Added
 
 - Introduce heroku build metrics support (https://github.com/heroku/buildpacks-ruby/pull/172)
@@ -23,5 +25,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial version of Ruby buildpack in Rust (https://github.com/heroku/buildpacks-ruby/pull/93)
 - Version 2.0.0 for the first release is not a typo. There was an issue in pack where a builder with the same name and version number would reuse artifacts left on image from [prior runs which caused issues](https://github.com/buildpacks/pack/issues/1322#issuecomment-1038241038). There were prior releases of `heroku/ruby` CNB from different sources that triggered this problem. To ensure no one would encounter that issue we developed and released using a version we know has not been used before. Version 2.0 was the first major version without a prior release of `heroku/ruby` CNB from any source.
 
-[unreleased]: https://github.com/heroku/buildpacks-ruby/compare/v2.0.1...HEAD
+[unreleased]: https://github.com/heroku/buildpacks-ruby/compare/v2.1.0...HEAD
+[2.1.0]: https://github.com/heroku/buildpacks-ruby/compare/v2.0.1...v2.1.0
 [2.0.1]: https://github.com/heroku/buildpacks-ruby/releases/tag/v2.0.1